requirements.txt
attrs
beautifulsoup4
coloredlogs
pandas
requests
